Redefine nativity
As the count of children younger than a certain age.

This metric is simpler and more stable than than the previous one.

+++ internal/ftop/renderer.go
@@ -19,7 +19,7 @@ const minHeight = 11
 type stats struct {
 	name     string
 	cpuTime  time.Duration
-	nativity float64
+	nativity uint
 	rssKb    int
 }
